프론트엔드 컴포넌트 리팩터링
