+++ src/components/Post.jsx
@@ -1,18 +1,27 @@
 /* eslint-disable no-nested-ternary */
 
+import { useEffect } from 'react';
+
 import styled from 'styled-components';
-import PostGame from './PostGame';
-import PostGameMembers from './PostGameMembers';
-import PostRegisterButton from './PostRegisterButton';
-import PostGameApplicants from './PostGameApplicants';
-import BackwardButton from './BackwardButton';
-import PostAuthorDetailAndImages from './ui/ComponentSectionContainer';
-import PostAuthor from './PostAuthor';
-import PostDetail from './PostDetail';
+
+import { useLocalStorage } from 'usehooks-ts';
+
+import usePostStore from '../hooks/usePostStore';
+import useGameStore from '../hooks/useGameStore';
+import usePlaceStore from '../hooks/usePlaceStore';
+import useRegisterStore from '../hooks/useRegisterStore';
+
 import Container from './ui/ComponentScreenContainer';
-import Button from './ui/PrimaryButton';
-import PostImages from './PostImages';
+import BackwardButton from './BackwardButton';
+
+import PostAuthorMenu from './PostAuthorMenu';
+import PostGame from './PostGame';
+import PostContent from './PostContent';
 import PostPlace from './PostPlace';
+import PostGameMembers from './PostGameMembers';
+import PostGameApplicants from './PostGameApplicants';
+import PostRegisterButton from './PostRegisterButton';
+import PostLoginGuidance from './PostLoginGuidance';
 
 const BackwardAndFunctions = styled.div`
   width: 100%;
@@ -20,10 +29,6 @@ const BackwardAndFunctions = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
-`;
-
-const Functions = styled.div`
-  
 `;
 
 const GameIsFull = styled.p`
@@ -36,73 +41,41 @@ const GameIsFull = styled.p`
   background-color: #A3A3A3;
 `;
 
-const LoginGuidance = styled.div`
-  font-size: .9em;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 1em;
-  padding: 1.25em;
-  border-radius: 5px;
-  background-color: #fff;
-`;
-
 export default function Post({
   loggedIn,
+  postId,
   navigateBackward,
   navigateLogin,
   navigateSelectTrialAccount,
-  post,
-  game,
-  place,
-  members,
-  applicants,
-  reconfirmDeletePost,
-  handleClickRegister,
-  reconfirmRegisterCancel,
-  reconfirmParticipateCancel,
-  handleClickAcceptRegister,
-  reconfirmRegisterReject,
-  registerError,
+  navigatePostsAfterDeleted,
 }) {
-  const onClickBackward = () => {
-    navigateBackward();
+  const [accessToken] = useLocalStorage('accessToken', '');
+
+  const postStore = usePostStore();
+  const gameStore = useGameStore();
+  const placeStore = usePlaceStore();
+  const registerStore = useRegisterStore();
+
+  const fetchData = async () => {
+    await postStore.fetchPost(postId);
+    const gameId = await gameStore.fetchGame(postId);
+    const { isAuthor } = postStore.post;
+    const { placeId } = gameStore.game;
+    await placeStore.fetchPlace(placeId);
+    await registerStore.fetchMembers(gameId);
+    if (isAuthor) {
+      await registerStore.fetchApplicants(gameId);
+    }
   };
 
-  const onClickLogin = () => {
-    navigateLogin();
-  };
+  useEffect(() => {
+    fetchData(postId);
+  }, [accessToken]);
 
-  const onClickSelectTrialAccount = () => {
-    navigateSelectTrialAccount();
-  };
+  const { post } = postStore;
+  const { game } = gameStore;
 
-  const onClickDeletePost = () => {
-    reconfirmDeletePost(post.id);
-  };
-
-  const onClickRegister = () => {
-    handleClickRegister(game.id);
-  };
-
-  const onClickRegisterCancel = () => {
-    reconfirmRegisterCancel(game.registerId);
-  };
-
-  const onClickParticipateCancel = () => {
-    reconfirmParticipateCancel(game.registerId);
-  };
-
-  const onClickAcceptRegister = (registerId) => {
-    handleClickAcceptRegister(registerId);
-  };
-
-  const onClickRejectRegister = (registerId) => {
-    reconfirmRegisterReject(registerId);
-  };
-
-  if (!post || !game || !members
+  if (!post || !game
     || Object.keys(post).length === 0
     || Object.keys(game).length === 0) {
     return (
@@ -114,109 +87,47 @@ export default function Post({
     <Container>
       <BackwardAndFunctions>
         <BackwardButton
-          onClick={onClickBackward}
+          onClick={navigateBackward}
         />
-        {post.isAuthor ? (
-          <Functions>
-            <Button
-              type="button"
-            >
-              수정하기
-            </Button>
-            <Button
-              type="button"
-              onClick={onClickDeletePost}
-            >
-              삭제하기
-            </Button>
-          </Functions>
-        ) : (
-          null
+        {post.isAuthor && (
+          <PostAuthorMenu
+            navigatePostsAfterDeleted={navigatePostsAfterDeleted}
+          />
         )}
       </BackwardAndFunctions>
-      {/* TODO: game.type을 name으로 바꿔야 함 */}
-      <PostGame
-        name={game.type}
-        date={game.date}
-        place={place.name}
-        currentMemberCount={game.currentMemberCount}
-        targetMemberCount={game.targetMemberCount}
-        hits={post.hits}
-      />
-      <PostAuthorDetailAndImages>
-        <PostAuthor
-          authorName={post.authorName}
-          authorPhoneNumber={post.authorPhoneNumber}
-          authorProfileImageUrl={post.authorProfileImageUrl}
-          authorMannerScore={post.authorMannerScore}
-        />
-        <PostDetail
-          detail={post.detail}
-        />
-        <PostImages
-          imageUrls={post.imageUrls}
-        />
-      </PostAuthorDetailAndImages>
-      <PostPlace
-        place={place}
-      />
-      <PostGameMembers
-        members={members}
-        isAuthor={post.isAuthor}
-        registerStatus={game.registerStatus}
-      />
+      <PostGame />
+      <PostContent />
+      <PostPlace />
+      <PostGameMembers />
       {loggedIn ? (
         post.isAuthor ? (
           <PostGameApplicants
-            applicants={applicants}
-            cannotAcceptRegister={(
-              game.currentMemberCount >= game.targetMemberCount
-            )}
-            onClickAcceptRegister={onClickAcceptRegister}
-            onClickRejectRegister={onClickRejectRegister}
+            fetchData={fetchData}
           />
         ) : (
-          game.registerStatus === 'processing' || game.registerStatus === 'accepted' ? (
-            <PostRegisterButton
-              registerStatus={game.registerStatus}
-              onClickRegister={onClickRegister}
-              onClickRegisterCancel={onClickRegisterCancel}
-              onClickParticipateCancel={onClickParticipateCancel}
-              registerError={registerError}
-            />
-          ) : (
-            game.currentMemberCount >= game.targetMemberCount ? (
-              <p>참가 정원이 모두 찼습니다.</p>
-            ) : (
+          game.registerStatus === 'processing'
+            || game.registerStatus === 'accepted' ? (
               <PostRegisterButton
-                registerStatus={game.registerStatus}
-                onClickRegister={onClickRegister}
-                onClickRegisterCancel={onClickRegisterCancel}
-                onClickParticipateCancel={onClickParticipateCancel}
-                registerError={registerError}
+                fetchData={fetchData}
               />
+            ) : (
+              game.currentMemberCount >= game.targetMemberCount ? (
+                <p>참가 정원이 모두 찼습니다.</p>
+              ) : (
+                <PostRegisterButton
+                  fetchData={fetchData}
+                />
+              )
             )
-          )
         )
       ) : (
         game.currentMemberCount >= game.targetMemberCount ? (
           <GameIsFull>참가 정원이 모두 찼습니다.</GameIsFull>
         ) : (
-          <LoginGuidance>
-            <p>참가를 신청하려면 로그인이 필요합니다.</p>
-            <Button
-              type="button"
-              onClick={onClickLogin}
-            >
-              로그인하기
-            </Button>
-            <Button
-              type="button"
-              onClick={onClickSelectTrialAccount}
-            >
-              체험 계정 선택하기
-            </Button>
-          </LoginGuidance>
+          <PostLoginGuidance
+            navigateLogin={navigateLogin}
+            navigateSelectTrialAccount={navigateSelectTrialAccount}
+          />
         )
       )}
     </Container>
