+++ src/components/PostDetail.jsx
@@ -1,4 +1,5 @@
 import styled from 'styled-components';
+import usePostStore from '../hooks/usePostStore';
 
 const Container = styled.section`
   font-size: .8em;
@@ -8,12 +9,22 @@ const Container = styled.section`
   line-height: 1.2;
 `;
 
-export default function PostDetail({
-  detail,
-}) {
+const Detail = styled.pre`
+  white-space: pre-wrap;
+  word-break: keep-all;
+  overflow: auto;
+`;
+
+export default function PostDetail() {
+  const postStore = usePostStore();
+
+  const { post } = postStore;
+
   return (
     <Container>
-      {detail}
+      <Detail>
+        {post.detail}
+      </Detail>
     </Container>
   );
 }
